DB Management Service/requirements.txt
PyJWT
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
requests==2.21.0
SQLAlchemy==1.3.1
Werkzeug==0.14.1
